Add row line with Moving Time
Because a GPX can loose the signal, or a stop can be done, the elapsed
time and the moving time may be different.
Change the Duration with the get_total_time, with is the time from start
to end of activity.
Add the Moving Time with the get_moving_time, previously displayed with
the Duration.